Address Raycast review feedback
