+++ extensions/workflowy/src/add-to-today.tsx
@@ -3,11 +3,7 @@ import { insertNode } from "./lib/api";
 import { insertNodeOptimistically } from "./lib/cache";
 import { getPreferences } from "./lib/preferences";
 
-interface Arguments {
-  text?: string;
-}
-
-export default async function Command(props: { arguments: Arguments }) {
+export default async function Command(props: { arguments: Arguments.AddToToday }) {
   const preferences = getPreferences();
   if (!preferences.apiKey) {
     await showToast({ style: Toast.Style.Failure, title: "Missing Workflowy API Key", message: "Open preferences to add it." });
@@ -15,7 +11,7 @@ export default async function Command(props: { arguments: Arguments }) {
     return;
   }
 
-  let text = props.arguments.text?.trim();
+  let text: string | undefined = props.arguments.text?.trim();
   if (!text) {
     try {
       text = (await getSelectedText()).trim();
